Fix errors and move trietxn to trie package
